requirements.txt
# 가상환경 설치: python -m venv .venv && .venv\Scripts\activate && pip install -r requirements.txt
# PDF 변환
pymupdf4llm>=0.0.20
# PowerPoint 변환 (1·2단계, 표는 Pandas 사용 시 tabulate 필요)
python-pptx>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
# 3단계 LLM 보정 (OpenAI API)
openai>=2.15.0
python-dotenv>=1.0.0
# 테스트
pytest>=7.0.0
